refactor TypeShell to implement pagination across project/type pages
